just saved

+++ ubxtool/between_redis_and_ubx.py
@@ -6,6 +6,7 @@ import threading
 import subprocess
 import json
 import re
+import os
 
 
 gpsd = None #seting the global variablecute
@@ -16,6 +17,7 @@ satellites = None
 redis_defaults = {
     'connection':'not connected',
     'rtk_source':'disabled',
+    'debug_level':0,
     'rtk':{
         'user':None,
         'password':None,
@@ -40,7 +42,8 @@ redis_defaults = {
             'status':None,
             'altHAE':None,
             'speed':None,
-            'eph':None
+            'eph':None,
+            'time':None
         },
         'SKY':{
             'hdop':None
@@ -69,21 +72,48 @@ class GpsPoller(threading.Thread):
         global gpsd
         global report
         while gps_thread.running:
-            report = gpsd.next() #this will continue to loop and grab EACH set of gpsd info to clear the buffer
+            try:
+                report = gpsd.next() #this will continue to loop and grab EACH set of gpsd info to clear the buffer
+            except:
+                continue
+            try:
+                if report['class'] == 'TPV':
+                    get_from_buffer('TPV')
+                if report['class'] == 'SKY':
+                    get_from_buffer('SKY')
+            except (KeyError, TypeError):
+                pass
+            time.sleep(0.5) #set to whatever
+
+
 
 class device_unplug_handler(threading.Thread):
     def __init__(self):
         threading.Thread.__init__(self)
+        self.running = True
     def run(self):
-        while gps_thread.running:
-            a = subprocess.run('ls /dev/serial/by-id',shell = True, stdout=subprocess.PIPE)
-            try:
-                re.search('usb-u-blox_AG_C099__ZED-F9P_DBTMNKT0-if00-port0', a.stdout.decode('utf-8')).group(0)
-                redis_client.set('connection', 'connected')
-            except AttributeError:
+        while device_unplug_handler_thread.running:
+            if os.path.isdir('/dev/serial/by-id') ==  True:
+                a = subprocess.run('ls /dev/serial/by-id',shell = True, stdout=subprocess.PIPE)
+                try:
+                    re.search('usb-u-blox_AG_C099__ZED-F9P_DBTMNKT0-if00-port0', a.stdout.decode('utf-8')).group(0)
+                    redis_client.set('connection', 'connected')
+                    time.sleep(1)
+                    gps_thread.running = True
+                    
+                    start_gpsd.run()
+
+
+                except AttributeError:
+                    gps_thread.running = False
+                    print('No devices connected')
+                    redis_client.set('connection','no connection')
+                    stop_gpsd.run()
+            else:
+                gps_thread.running = False
                 print('No devices connected')
                 redis_client.set('connection','no connection')
-                #restart_gpsd()
+                stop_gpsd.run()
             time.sleep(0.2)
 
 class ubx_to_redis(threading.Thread):
@@ -102,7 +132,13 @@ class ubx_to_redis(threading.Thread):
                 if int(c) != redis_defaults['ubxtool'][item]:
                     print(c, '<--!=-->',redis_defaults['ubxtool'][item])
                     app = run('ubxtool -P 27.12 -z {},{}'.format(item, redis_defaults['ubxtool'][item]))
-                    if re.search('UBX-ACK-NAK', )
+                    try:
+                        if re.findall('UBX-ACK-\w*', app.decode('utf-8'))[0] == 'UBX-ACK-NAK':
+                            redis_defaults['ubxtool'][item] = int(c)
+                            redis_client.set(item, c)
+                    except IndexError:
+                        redis_defaults['ubxtool'][item] = int(c) 
+                        redis_client.set(item, c)
     def ubx_get_item(self, item):
         return 'ubxtool -P 27.12 -g {}'.format(item)
 
@@ -113,49 +149,85 @@ class redis_get(threading.Thread):
         while gps_thread.running:
             for item in list(redis_defaults['ubxtool'].keys()):
                 if redis_client.exists(item) != 0:
-                    redis_defaults['ubxtool'][item] = int(redis_client.get(item))
+                    try:
+                        redis_defaults['ubxtool'][item] = int(redis_client.get(item))
+                    except ValueError:
+                        redis_client.set(item, redis_defaults['ubxtool'][item])
                 elif redis_client.exists(item) == 0:
                     redis_client.set(item,redis_defaults['ubxtool'][item])
             time.sleep(2)
+            if redis_client.exists('debug_level') != 0:
+                try:
+                    redis_defaults['debug_level'] = int(redis_client.get('debug_level'))
+                except ValueError:
+                    redis_client.set(item, redis_defaults['debug_level'])
+            else:
+                redis_client.set('debug_level',redis_defaults['debug_level'])
 
 def run(command):
-    #print(command)
     p = subprocess.Popen(command, shell = True, stdout = subprocess.PIPE)
     (output, err) = p.communicate()
     p_status =  p.wait()
+    if redis_defaults['debug_level'] == 1:
+        print(command)
+        redis_client.set('log', output.decode('utf-8'))
     #print(output.decode('utf-8'))
     return output
+def get(data, key):
+    try:
+        result = getattr(data, key, 'Unknown')
+    except AttributeError:
+        return 'Unknown'
+    return result
 
 def get_from_buffer(type):
     global satellites
     if type == "TPV":
         for field in list(redis_defaults['gpsd']['TPV'].keys()):
-            result = getattr(report, field, 'Unknown')
+            
+            result = get(report, field, 'Unknown')
             #print(field, ": ",result)
             redis_client.set(field,str(result))
     elif type == "SKY":
         for field in list(redis_defaults['gpsd']['SKY'].keys()):
-            result = getattr(report, field, 'Unknown')
+            result = get(report, field, 'Unknown')
             #print(field, ": ",result)
             redis_client.set(field,str(result))
-        satellites = getattr(report, 'satellites')
+        satellites = get(report, 'satellites')
         sat_used = 0
         for sat in satellites:
             if sat['used']==True:
                 sat_used+=1
         redis_client.set('sat_used',str(sat_used))
 
-def restart_gpsd():
-    print('Stopping gpsd')
-    subprocess.run('echo andrew | sudo -S systemctl stop gpsd',\
-            shell = True, check =True)
-    time.sleep(5)
-    print('Starting gpsd')
-    subprocess.run('echo andrew | sudo -S systemctl start gpsd',\
-            shell = True, check =True)
-
+class start_gpsd_class():
+    def __init__(self):
+        self.counter = 1
+    def run(self):
+        if self.counter > 0:
+            print('Starting gpsd')
+            time.sleep(2)
+            subprocess.run('echo andrew | sudo -S systemctl start gpsd',\
+                    shell = True, check =True)
+            self.counter = 0
+            stop_gpsd.counter = 1
+class stop_gpsd_class():
+    def __init__(self):
+        self.counter = 1
+    def run(self):
+        if self.counter > 0:
+            print('Stoping gpsd')
+            time.sleep(2)
+            subprocess.run('echo andrew | sudo -S systemctl stop gpsd',\
+                    shell = True, check =True)
+            self.counter = 0
+            start_gpsd.counter = 1
+start_gpsd = start_gpsd_class()
+stop_gpsd = stop_gpsd_class()
 
 if __name__ == '__main__':
+    stop_gpsd = stop_gpsd_class()
+    start_gpsd = start_gpsd_class()
     redis_client = redis.Redis(**redis_connection)
     redis_get_thread = redis_get()
     gps_thread = GpsPoller() # create the thread
@@ -166,17 +238,7 @@ if __name__ == '__main__':
         gps_thread.start() # start it up
         device_unplug_handler_thread.start()
         ubx_to_redis_thread.start()
-        while True:
-        #It may take a second or two to get good data
-        #print gpsd.fix.latitude,', ',gpsd.fix.longitude,'  Time: ',gpsd.utc
-            try:
-                if report['class'] == 'TPV':
-                    get_from_buffer('TPV')
-                if report['class'] == 'SKY':
-                    get_from_buffer('SKY')
-            except (KeyError, TypeError):
-                pass
-            time.sleep(0.5) #set to whatever
+
 
     except (KeyboardInterrupt, SystemExit): #when you press ctrl+c
         print("\nKilling Thread...")
